REFACATOR function story subject behaviour calls and use env var
