+++ src/app/core/services/plot/plot.service.ts
@@ -10,6 +10,8 @@ import * as JSON5 from 'json5';
 import * as uuid from "uuid";
 
 // TODO: move function and variables still used to stories-service.
+// TODO: Rename this Service. Something like "NodeService" as it may be handling Graph Node related changes
+// Project is convoluted. Must fix
 @Injectable({
   providedIn: 'root'
 })
@@ -46,24 +48,24 @@ export class PlotService {
     private http: HttpClient,
   ) { }
 
-  GetStory(): Observable<Plot[]> {
-    return this.http.get<PlotContent>(this.storyJSON)
-      // Error  
-      .pipe(
-        catchError((error: Error, caught: Observable<any>) => {
-          console.warn("Getting Stories Error:", error.message);
-          throw new Error(error.message);
-          return ([])
-        })
-      )
-      // Request 
-      .pipe(
-        tap((res: Plot[]) => {
-          console.info("Getting Stories map", res);
-          return res;
-        })
-      )
-  }
+  // GetStory(): Observable<Plot[]> {
+  //   return this.http.get<PlotContent>(this.storyJSON)
+  //     // Error  
+  //     .pipe(
+  //       catchError((error: Error, caught: Observable<any>) => {
+  //         console.warn("Getting Stories Error:", error.message);
+  //         throw new Error(error.message);
+  //         return ([])
+  //       })
+  //     )
+  //     // Request 
+  //     .pipe(
+  //       tap((res: Plot[]) => {
+  //         console.info("Getting Stories map", res);
+  //         return res;
+  //       })
+  //     )
+  // }
 
   /**
    * Request to get all Stories. These are false data. Stored in `/data`.
@@ -73,31 +75,32 @@ export class PlotService {
   UpdateStoryBehavior(id: string): void {
     console.log('function call update story behavior.')
     // update store
-    this.GetStory().subscribe((response: Plot[]) => {
-      (response && response.length) ? this.storyBehaviorSubject.next(response.find((ob: Plot) => ob.id === id)) : null
-    });
+    // this.GetStory().subscribe((response: Plot[]) => {
+    //   (response && response.length) ? this.storyBehaviorSubject.next(response.find((ob: Plot) => ob.id === id)) : null
+    // });
   }
 
-  /**
-   * Create a brand new story graph.
-   * @param id uuid of Story 
-   */
-  createStoryGraph(id: string) {
-    const newStory: Plot = {
-      id,
-      description: "",
-      title: "-",
-      content: {
-        id: uuid.v4(),
-        name: "brand new",
-        description: "-",
-        children: undefined,
-        graphics: undefined,
-        characters: undefined,
-      }
-    }
-    this.storyBehaviorSubject.next(newStory);
-  }
+  // /**
+  //  * Create a brand new story graph.
+  //  * @param id uuid of Story 
+  //  */
+  // createNewStoryGraph(id: string) {
+  //   const story: Plot = {
+  //     id,
+  //     description: "Description text needed",
+  //     title: "Title of Story",
+  //     content: {
+  //       id: uuid.v4(),
+  //       name: "Initial Content for story",
+  //       description: "-",
+  //       children: undefined,
+  //       graphics: undefined,
+  //       characters: undefined,
+  //     }
+  //   }
+
+  //   this.storyBehaviorSubject.next(story);
+  // }
 
   selectInstance({ instance, parentInstanceId }: { instance: PlotContent, parentInstanceId?: string }) {
     this.instanceEditSubject.next({ instance, parentInstanceId });
